refactor(orchestrator): extract evidence helpers
