+++ src/ouroboros/orchestrator/parallel_executor.py
@@ -32,11 +32,7 @@ from dataclasses import replace
 from datetime import UTC, datetime
 import json
 import os
-from pathlib import Path
-import platform
 import re
-import shlex
-import subprocess
 import time
 from typing import TYPE_CHECKING, Any
 
@@ -70,6 +66,143 @@ from ouroboros.orchestrator.effort_routing import resolve_execute_effort
 from ouroboros.orchestrator.events import (
     create_ac_stall_detected_event,
     create_heartbeat_event,
+)
+from ouroboros.orchestrator.evidence.ac_classification import (  # noqa: F401
+    _CODE_IMPLEMENTATION_ACTION_RE,
+    _CODE_MUTATION_ACTION_RE,
+    _CODE_WORK_SIGNAL_RE,
+    _DOC_ONLY_ACTION_RE,
+    _DOC_ONLY_TARGET_RE,
+    _DOCS_TEST_REFERENCE_RE,
+    _EXISTING_VALIDATION_RE,
+    _NO_MUTATION_VALIDATION_RE,
+    _TEST_MUTATION_WORK_RE,
+    _TEST_WORK_RE,
+    _VALIDATION_ONLY_ACTION_RE,
+    _VALIDATION_ONLY_TEST_SIGNAL_RE,
+    _effective_evidence_schema_for_ac,
+    _has_mixed_code_and_documentation_work,
+    _has_mixed_test_and_documentation_work,
+    _has_mixed_validation_and_documentation_work,
+    _is_documentation_only_ac,
+    _is_validation_only_ac,
+    _out_of_scope_evidence_fields_for_ac,
+    _out_of_scope_evidence_values_for_ac,
+    _profile_with_evidence_schema,
+    _scoped_evidence_record_for_ac,
+)
+from ouroboros.orchestrator.evidence.claims import (  # noqa: F401
+    _bash_command_mutates_file_reference,
+    _file_claim_matches_runtime_path,
+    _file_reference_pattern,
+    _runtime_command_value_to_text,
+    _runtime_message_command_values,
+    _runtime_message_file_path_values,
+    _runtime_message_file_proof_text,
+    _runtime_message_has_following_success,
+    _runtime_message_has_success_evidence,
+    _runtime_message_has_success_signal,
+    _runtime_message_search_text,
+    _runtime_message_supports_command_claim,
+    _runtime_message_supports_file_reference,
+    _runtime_messages_have_masked_test_command_form,
+    _runtime_messages_support_claim,
+    _runtime_messages_support_command_claim,
+    _runtime_messages_support_file_claim,
+    _runtime_support_messages_for_field,
+    _text_supports_file_mutation_reference,
+    _workspace_relative_file_claim,
+)
+from ouroboros.orchestrator.evidence.common import (  # noqa: F401
+    _MAX_LEAF_RESULT_CHARS,
+    _flatten_evidence_values,
+    _normalize_command,
+    _normalize_exact_command,
+    _normalized_evidence_text,
+    _truncate_text,
+)
+from ouroboros.orchestrator.evidence.formatting import (  # noqa: F401
+    _build_governed_parent_summary,
+    _extract_leaf_evidence_lines,
+    _render_ac_section,
+    _subtask_event_label,
+)
+from ouroboros.orchestrator.evidence.runtime_metadata import (  # noqa: F401
+    _AC_RUNTIME_OWNERSHIP_METADATA_KEYS,
+    _AC_RUNTIME_RESUME_METADATA_KEYS,
+    _AC_RUNTIME_SCOPE_METADATA_KEYS,
+    _NON_REUSABLE_RUNTIME_EVENT_TYPES,
+    _REUSABLE_RUNTIME_EVENT_TYPES,
+    _SIBLING_HEADLINE_CHARS,
+    _STALL_SENTINEL,
+    HEARTBEAT_INTERVAL_SECONDS,
+    MAX_STALL_RETRIES,
+    STALL_TIMEOUT_SECONDS,
+    _SiblingACRef,
+)
+from ouroboros.orchestrator.evidence.shell_parsing import (  # noqa: F401
+    _OUTPUT_FILTER_COMMANDS,
+    _TRAILING_REDIRECT_RE,
+    _has_gradle_or_maven_test_skip,
+    _has_trailing_output_filter_pipeline,
+    _is_env_assignment,
+    _is_pipefail_parts,
+    _is_pipefail_preamble,
+    _is_safe_test_command_preamble,
+    _looks_like_test_command,
+    _looks_like_unittest_command,
+    _normalized_command_claim_aliases,
+    _normalized_shell_words_text,
+    _output_filter_pipeline_is_pipefail_protected,
+    _runtime_command_evidence_aliases,
+    _segments_after_safe_shell_preamble,
+    _segments_after_safe_shell_preamble_with_pipefail,
+    _shell_command_body,
+    _single_command_after_safe_shell_preamble,
+    _single_exact_command_after_safe_shell_preamble,
+    _strip_command_output_plumbing,
+    _strip_env_prefix,
+    _test_command_invocation,
+    _test_command_invocation_allowing_output_plumbing,
+    _test_invocation_from_prefix,
+    _test_invocation_from_shell_body,
+    _unittest_command_invocation,
+    _uses_pipefail,
+)
+from ouroboros.orchestrator.evidence.system import (  # noqa: F401
+    _MEMORY_CHECK_INTERVAL_SECONDS,
+    _MEMORY_WAIT_MAX_SECONDS,
+    _MIN_FREE_MEMORY_GB,
+    _get_available_memory_gb,
+)
+from ouroboros.orchestrator.evidence.test_detection import (  # noqa: F401
+    _claim_contains_command_success_summary,
+    _claim_summary_matches_runtime_chunk,
+    _is_tool_result_message,
+    _message_contains_test_success,
+    _runtime_message_test_proof_text,
+    _runtime_messages_have_masked_test_command_for_test_claim,
+    _runtime_messages_support_test_claim,
+    _successful_runtime_test_commands,
+    _test_claim_file_part,
+    _test_command_targets_claim,
+    _text_contains_test_success,
+    _text_contains_unittest_success,
+)
+from ouroboros.orchestrator.evidence.typed_evidence import (  # noqa: F401
+    _add_runtime_command_evidence,
+    _complete_sibling_acs_from_evidence,
+    _criterion_inline_code_values,
+    _criterion_is_exact_command_pass_ac,
+    _criterion_is_exact_command_run_ac,
+    _criterion_is_exact_file_presence_ac,
+    _criterion_satisfied_by_evidence,
+    _evidence_values_from_result,
+    _typed_evidence_is_usable_for_sibling_reconciliation,
+    _typed_file_evidence_proves_current_existence,
+)
+from ouroboros.orchestrator.evidence.verification import (
+    _verify_atomic_evidence_against_runtime_messages,
 )
 from ouroboros.orchestrator.evidence_schema import (
     EvidenceError,
@@ -107,7 +240,7 @@ from ouroboros.orchestrator.policy import (
     PolicySessionRole,
     evaluate_capability_policy,
 )
-from ouroboros.orchestrator.profile_loader import EvidenceSchema, ExecutionProfile
+from ouroboros.orchestrator.profile_loader import ExecutionProfile
 from ouroboros.orchestrator.rate_limit import (
     RateLimitBackoff,
     RateLimitGate,
@@ -147,2152 +280,6 @@ MIN_SUB_ACS = 2
 MAX_SUB_ACS = 5
 DECOMPOSITION_TIMEOUT_SECONDS = 60.0
 _IMPLEMENTATION_SESSION_KIND = "implementation_session"
-
-
-_DOC_ONLY_TARGET_RE = re.compile(
-    r"\b(readme(?:\.md)?|docs?/|docs?\.[a-z0-9_-]+|documentation|guide|manual|changelog)\b",
-    re.IGNORECASE,
-)
-_DOC_ONLY_ACTION_RE = re.compile(
-    r"\b(document|describe|explain|add|update|create|fix|write|improve)\b",
-    re.IGNORECASE,
-)
-_CODE_IMPLEMENTATION_ACTION_RE = re.compile(
-    r"\b(implement|build|develop|ship)\b",
-    re.IGNORECASE,
-)
-_CODE_MUTATION_ACTION_RE = re.compile(
-    r"\b(add(?:ing)?|fix(?:ing)?|create|creating|update|updating|change|changing|modify|modifying|refactor(?:ing)?|repair(?:ing)?)\b",
-    re.IGNORECASE,
-)
-_CODE_WORK_SIGNAL_RE = re.compile(
-    r"("
-    r"\b[a-zA-Z_][a-zA-Z0-9_]*\s*\("
-    r"|"
-    r"\.(?:py|pyi|js|jsx|ts|tsx|go|rs|java|kt|c|cc|cpp|h|hpp|swift|rb|php|sh|zsh|fish)\b"
-    r"|"
-    r"\b(parser|function|module|api|endpoint|class|method|cli\s+flag|flag|command|"
-    r"bug|runtime|workflow|validator|validation|implementation)\b"
-    r")",
-    re.IGNORECASE,
-)
-_TEST_WORK_RE = re.compile(
-    r"("
-    r"\b(?:run|execute|pass|validate|verify)\b.{0,40}\b(?:pytest|tests?|unit\s+tests?|integration\s+tests?)\b"
-    r"|"
-    r"\b(?:add|write|create|implement|fix|update)\b.{0,40}\b"
-    r"(?:tests?|unit\s+tests?|integration\s+tests?)\b"
-    r"(?!\s+(?:guide|docs?|documentation|setup))"
-    r"|"
-    r"\b(?:pytest|tests_passed|test\s+command)\b"
-    r")",
-    re.IGNORECASE,
-)
-_TEST_MUTATION_WORK_RE = re.compile(
-    r"("
-    r"\b(?:add|write|create|implement|fix|update|extend|expand)\b.{0,60}\b"
-    r"(?:tests?|unit\s+tests?|integration\s+tests?|coverage|test_[\w.-]+\.py)\b"
-    r"|"
-    r"\b(?:tests?|unit\s+tests?|integration\s+tests?|test_[\w.-]+\.py)\b.{0,60}\b"
-    r"(?:cover|coverage)\b"
-    r"|"
-    r"\bcheck(?:ed)?\s+(?:(?:the|existing|current|new|added|updated)\s+){0,3}"
-    r"(?:tests?|unit\s+tests?|integration\s+tests?|test_[\w.-]+\.py)"
-    r"\s+into\b"
-    r"|"
-    r"\bcheck\s+in\b.{0,60}\b"
-    r"(?:tests?|unit\s+tests?|integration\s+tests?|test_[\w.-]+\.py)\b"
-    r")",
-    re.IGNORECASE,
-)
-_DOCS_TEST_REFERENCE_RE = re.compile(
-    r"("
-    r"\b(?:document(?:ing)?|guide|manual|instructions?|usage|setup|how\s+to|verification)\b"
-    r".{0,100}\b(?:test\s+command|python\s+-m\s+unittest|pytest|tests?|unit\s+tests?|test\s+setup)\b"
-    r"|"
-    r"\b(?:test\s+command|python\s+-m\s+unittest|pytest|tests?|unit\s+tests?|test\s+setup)\b"
-    r".{0,100}\b(?:document(?:ing)?|guide|manual|instructions?|usage|setup|how\s+to|verification)\b"
-    r")",
-    re.IGNORECASE,
-)
-_NO_MUTATION_VALIDATION_RE = re.compile(
-    r"("
-    r"\bwithout\s+(?:modifying|changing|editing|writing|updating|touching)\b"
-    r"|"
-    r"\bwith\s+no\s+(?:file|code)?\s*(?:modifications?|changes?|edits?|updates?)\b"
-    r"|"
-    r"\bno\s+(?:file|code)?\s*(?:modifications?|changes?|edits?|updates?)\b"
-    r"|"
-    r"\bdo\s+not\s+(?:modify|change|edit|write|update|touch)\b"
-    r")",
-    re.IGNORECASE,
-)
-_EXISTING_VALIDATION_RE = re.compile(
-    r"\b(?:existing|current|already(?:-|\s+)?satisfied|already(?:-|\s+)?implemented)\b",
-    re.IGNORECASE,
-)
-_VALIDATION_ONLY_ACTION_RE = re.compile(
-    r"\b(?:run|execute|pass|validate|verify|ensure|confirm|check)\b",
-    re.IGNORECASE,
-)
-_VALIDATION_ONLY_TEST_SIGNAL_RE = re.compile(
-    r"\b(?:pytest|unit\s+tests?|integration\s+tests?|tests?|test suite|"
-    r"test_[\w.-]+\.py|python\s+-m\s+unittest)\b",
-    re.IGNORECASE,
-)
-
-
-def _has_mixed_code_and_documentation_work(ac_content: str) -> bool:
-    """Return True when one AC appears to combine code mutation and docs work."""
-    for connector in re.finditer(
-        r"\b(?:and|then|while|plus)\b|[,;:]",
-        ac_content,
-        re.IGNORECASE,
-    ):
-        before = ac_content[: connector.start()]
-        after = ac_content[connector.end() :]
-        before_has_docs = bool(_DOC_ONLY_TARGET_RE.search(before))
-        after_has_docs = bool(_DOC_ONLY_TARGET_RE.search(after))
-        before_has_code_work = bool(
-            _CODE_MUTATION_ACTION_RE.search(before) and _CODE_WORK_SIGNAL_RE.search(before)
-        )
-        after_has_code_work = bool(
-            _CODE_MUTATION_ACTION_RE.search(after) and _CODE_WORK_SIGNAL_RE.search(after)
-        )
-        if after_has_docs and not before_has_docs and before_has_code_work:
-            return True
-        if before_has_docs and not after_has_docs and after_has_code_work:
-            return True
-    return False
-
-
-def _has_mixed_test_and_documentation_work(ac_content: str) -> bool:
-    """Return True when one AC appears to combine test mutation and docs work."""
-    for connector in re.finditer(
-        r"\b(?:and|then|while|plus)\b|[,;:]",
-        ac_content,
-        re.IGNORECASE,
-    ):
-        before = ac_content[: connector.start()]
-        after = ac_content[connector.end() :]
-        before_has_docs = bool(_DOC_ONLY_TARGET_RE.search(before))
-        after_has_docs = bool(_DOC_ONLY_TARGET_RE.search(after))
-        before_has_test_work = bool(_TEST_MUTATION_WORK_RE.search(before))
-        after_has_test_work = bool(_TEST_MUTATION_WORK_RE.search(after))
-        if after_has_docs and not before_has_docs and before_has_test_work:
-            return True
-        if before_has_docs and not after_has_docs and after_has_test_work:
-            return True
-    return False
-
-
-def _has_mixed_validation_and_documentation_work(ac_content: str) -> bool:
-    """Return True when one AC appears to combine docs work and test execution."""
-    for connector in re.finditer(
-        r"\b(?:and|then|while|plus)\b|[,;:]",
-        ac_content,
-        re.IGNORECASE,
-    ):
-        before = ac_content[: connector.start()]
-        after = ac_content[connector.end() :]
-        before_has_docs = bool(_DOC_ONLY_TARGET_RE.search(before))
-        after_has_docs = bool(_DOC_ONLY_TARGET_RE.search(after))
-        before_has_validation = bool(
-            _VALIDATION_ONLY_ACTION_RE.search(before)
-            and _VALIDATION_ONLY_TEST_SIGNAL_RE.search(before)
-        )
-        after_has_validation = bool(
-            _VALIDATION_ONLY_ACTION_RE.search(after)
-            and _VALIDATION_ONLY_TEST_SIGNAL_RE.search(after)
-        )
-        if after_has_docs and not before_has_docs and before_has_validation:
-            return True
-        if before_has_docs and not after_has_docs and after_has_validation:
-            return True
-    return False
-
-
-def _is_documentation_only_ac(ac_content: str) -> bool:
-    """Return True when an AC asks only for documentation/README work.
-
-    The code profile normally requires runnable test evidence. Normal usage can
-    still include a final docs-only AC (for example README usage examples after
-    code ACs already passed). Such an AC should be verified by docs evidence
-    from the current runtime session, not by re-claiming prior code test IDs.
-    """
-    normalized = " ".join(ac_content.split())
-    if not normalized:
-        return False
-    has_docs_target = bool(_DOC_ONLY_TARGET_RE.search(normalized))
-    has_docs_action = bool(_DOC_ONLY_ACTION_RE.search(normalized))
-    documents_test_reference = (
-        has_docs_target and has_docs_action and bool(_DOCS_TEST_REFERENCE_RE.search(normalized))
-    )
-    if documents_test_reference and _has_mixed_validation_and_documentation_work(normalized):
-        return False
-    if _TEST_MUTATION_WORK_RE.search(normalized) and (
-        not documents_test_reference or _has_mixed_test_and_documentation_work(normalized)
-    ):
-        return False
-    if _TEST_WORK_RE.search(normalized) and not documents_test_reference:
-        return False
-    if _CODE_IMPLEMENTATION_ACTION_RE.search(normalized):
-        return False
-    if _has_mixed_code_and_documentation_work(normalized):
-        return False
-    if (
-        re.search(r"\bdocumentation\b", normalized, re.IGNORECASE)
-        and _CODE_MUTATION_ACTION_RE.search(normalized)
-        and _CODE_WORK_SIGNAL_RE.search(normalized)
-        and not re.search(
-            r"\b(readme(?:\.md)?|docs?/|docs?\.[a-z0-9_-]+|guide|manual|changelog)\b",
-            normalized,
-            re.IGNORECASE,
-        )
-    ):
-        return False
-    return has_docs_target and has_docs_action
-
-
-def _is_validation_only_ac(ac_content: str) -> bool:
-    """Return True when an AC asks only to run or verify tests.
-
-    Test-writing ACs still require ``files_touched``; validation-only ACs are
-    allowed to prove completion with command/test evidence and no file mutation.
-    """
-    normalized = " ".join(ac_content.split())
-    if not normalized:
-        return False
-    if _is_documentation_only_ac(normalized):
-        return False
-    if _DOC_ONLY_TARGET_RE.search(normalized) and _DOC_ONLY_ACTION_RE.search(normalized):
-        return False
-    stripped_no_mutation_terms = _NO_MUTATION_VALIDATION_RE.sub("", normalized)
-    if _TEST_MUTATION_WORK_RE.search(normalized):
-        return False
-    if _CODE_MUTATION_ACTION_RE.search(stripped_no_mutation_terms) and _CODE_WORK_SIGNAL_RE.search(
-        stripped_no_mutation_terms
-    ):
-        return False
-    if _CODE_IMPLEMENTATION_ACTION_RE.search(stripped_no_mutation_terms):
-        return False
-    if _has_mixed_code_and_documentation_work(stripped_no_mutation_terms):
-        return False
-    if (
-        (
-            _NO_MUTATION_VALIDATION_RE.search(normalized)
-            or _EXISTING_VALIDATION_RE.search(normalized)
-        )
-        and _VALIDATION_ONLY_ACTION_RE.search(normalized)
-        and _VALIDATION_ONLY_TEST_SIGNAL_RE.search(normalized)
-    ):
-        return True
-    return bool(_VALIDATION_ONLY_ACTION_RE.search(normalized)) and bool(
-        _VALIDATION_ONLY_TEST_SIGNAL_RE.search(normalized)
-    )
-
-
-def _effective_evidence_schema_for_ac(
-    profile: ExecutionProfile,
-    ac_content: str,
-) -> EvidenceSchema:
-    """Return the active evidence schema for one atomic AC dispatch."""
-    schema = profile.evidence_schema
-    if _is_validation_only_ac(ac_content) and "files_touched" in schema.required:
-        required = tuple(field for field in schema.required if field != "files_touched")
-        rejected_if = tuple(
-            expr for expr in schema.rejected_if if not expr.strip().startswith("files_touched")
-        )
-        return EvidenceSchema(required=required, rejected_if=rejected_if)
-    if not _is_documentation_only_ac(ac_content) or "tests_passed" not in schema.required:
-        return schema
-    required = tuple(field for field in schema.required if field != "tests_passed")
-    rejected_if = tuple(
-        expr for expr in schema.rejected_if if not expr.strip().startswith("tests_passed")
-    )
-    return EvidenceSchema(required=required, rejected_if=rejected_if)
-
-
-def _out_of_scope_evidence_fields_for_ac(
-    profile: ExecutionProfile,
-    ac_content: str,
-    record: EvidenceRecord | None,
-) -> tuple[str, ...]:
-    """Return non-empty evidence fields excluded by the AC-specific schema."""
-    if record is None:
-        return ()
-    effective_schema = _effective_evidence_schema_for_ac(profile, ac_content)
-    required_fields = set(effective_schema.required)
-    return tuple(
-        field
-        for field in profile.evidence_schema.required
-        if field not in required_fields and _flatten_evidence_values(record.get(field))
-    )
-
-
-def _out_of_scope_evidence_values_for_ac(
-    profile: ExecutionProfile,
-    ac_content: str,
-    record: EvidenceRecord | None,
-) -> dict[str, Any]:
-    """Return out-of-scope evidence values retained for audit metadata only."""
-    if record is None:
-        return {}
-    fields = _out_of_scope_evidence_fields_for_ac(profile, ac_content, record)
-    return {field: record.data[field] for field in fields if field in record.data}
-
-
-def _scoped_evidence_record_for_ac(
-    profile: ExecutionProfile,
-    ac_content: str,
-    record: EvidenceRecord,
-) -> EvidenceRecord:
-    """Return only evidence fields inside the AC-specific schema."""
-    effective_schema = _effective_evidence_schema_for_ac(profile, ac_content)
-    allowed_fields = set(effective_schema.required)
-    return EvidenceRecord(
-        data={field: value for field, value in record.data.items() if field in allowed_fields},
-        source=record.source,
-    )
-
-
-def _profile_with_evidence_schema(
-    profile: ExecutionProfile,
-    schema: EvidenceSchema,
-) -> ExecutionProfile:
-    """Return a shallow profile copy using an AC-specific evidence schema."""
-    required_fields = set(schema.required)
-    must_produce = tuple(field for field in profile.must_produce if field in required_fields)
-    return profile.model_copy(update={"evidence_schema": schema, "must_produce": must_produce})
-
-
-def _subtask_event_label(content: str, *, max_length: int = 50) -> str:
-    """Return compact display text without losing full event content."""
-    normalized = " ".join(content.split())
-    if len(normalized) <= max_length:
-        return normalized
-    return normalized[:max_length]
-
-
-def _flatten_evidence_values(value: object) -> tuple[str, ...]:
-    """Return concrete string claims from a typed evidence field."""
-    if value is None:
-        return ()
-    if isinstance(value, str):
-        stripped = value.strip()
-        return (stripped,) if stripped else ()
-    if isinstance(value, (int, float, bool)):
-        return (str(value),)
-    if isinstance(value, dict):
-        flattened: list[str] = []
-        for item in value.values():
-            flattened.extend(_flatten_evidence_values(item))
-        return tuple(flattened)
-    if isinstance(value, (list, tuple, set)):
-        flattened_sequence: list[str] = []
-        for item in value:
-            flattened_sequence.extend(_flatten_evidence_values(item))
-        return tuple(flattened_sequence)
-    return (str(value),)
-
-
-def _runtime_message_search_text(message: AgentMessage) -> str:
-    """Build searchable transcript text for one non-final runtime message."""
-    parts: list[str] = [message.content]
-    if message.tool_name:
-        parts.append(message.tool_name)
-    tool_input = message.data.get("tool_input")
-    if isinstance(tool_input, dict):
-        parts.extend(str(value) for value in tool_input.values() if value is not None)
-    parts.extend(_flatten_evidence_values(message.data))
-    return "\n".join(parts).lower()
-
-
-def _runtime_message_file_path_values(message: AgentMessage) -> tuple[str, ...]:
-    """Return explicit file path values carried by a runtime message.
-
-    Codex/OpenCode file-change events may report absolute workspace paths while
-    typed evidence should normally claim workspace-relative paths. Keep this
-    structured path extraction separate from broad text search so read-only text
-    mentions still cannot prove ``files_touched``.
-    """
-    path_keys = {
-        "file_path",
-        "filepath",
-        "filePath",
-        "notebook_path",
-        "notebookPath",
-        "path",
-        "target_file",
-        "targetFile",
-    }
-    values: list[str] = []
-
-    def visit(value: object) -> None:
-        if isinstance(value, dict):
-            for key, child in value.items():
-                if key in path_keys and isinstance(child, str) and child.strip():
-                    values.append(child.strip())
-                else:
-                    visit(child)
-        elif isinstance(value, (list, tuple)):
-            for child in value:
-                visit(child)
-
-    for container_key in ("tool_input", "input", "arguments", "args"):
-        visit(message.data.get(container_key))
-    return tuple(values)
-
-
-def _runtime_message_command_values(message: AgentMessage) -> tuple[str, ...]:
-    """Return explicit command strings carried by a runtime message.
-
-    Runtime adapters normalize shell calls slightly differently.  Codex-like
-    events usually expose ``tool_input.command``; Goose may expose ``cmd`` or a
-    list argv form.  Keep extraction structured, not prose-based, so command
-    evidence does not fall back to arbitrary assistant text.
-    """
-    values: list[str] = []
-    for container_key in ("tool_input", "input", "arguments", "args"):
-        container = message.data.get(container_key)
-        if not isinstance(container, dict):
-            continue
-        for command_key in ("command", "cmd", "command_line"):
-            command = container.get(command_key)
-            normalized = _runtime_command_value_to_text(command)
-            if normalized and normalized not in values:
-                values.append(normalized)
-    return tuple(values)
-
-
-def _runtime_command_value_to_text(value: object) -> str | None:
-    """Normalize a structured runtime command value into shell text."""
-    if isinstance(value, str) and value.strip():
-        return value.strip()
-    if isinstance(value, list) and value:
-        return shlex.join(str(part) for part in value)
-    return None
-
-
-def _file_claim_matches_runtime_path(
-    claim: str,
-    runtime_path: str,
-    *,
-    task_cwd: str | None,
-) -> bool:
-    """Return True when a claimed workspace path matches a runtime path value."""
-    claim_path = Path(claim.strip())
-    if not claim_path or claim_path.is_absolute() or ".." in claim_path.parts:
-        return False
-
-    runtime_path = runtime_path.strip()
-    if not runtime_path:
-        return False
-
-    runtime_candidate = Path(runtime_path)
-    if task_cwd is not None:
-        base = Path(task_cwd).resolve()
-        claimed_absolute = (base / claim_path).resolve()
-        runtime_absolute = (
-            runtime_candidate if runtime_candidate.is_absolute() else base / runtime_candidate
-        ).resolve()
-        try:
-            claimed_absolute.relative_to(base)
-            runtime_absolute.relative_to(base)
-        except ValueError:
-            return False
-        return runtime_absolute == claimed_absolute
-
-    normalized_claim = claim_path.as_posix().lower()
-    normalized_runtime = runtime_candidate.as_posix().lower()
-    return normalized_runtime == normalized_claim or normalized_runtime.endswith(
-        "/" + normalized_claim
-    )
-
-
-def _workspace_relative_file_claim(value: str, *, task_cwd: str | None) -> str | None:
-    """Normalize a files_touched claim to a workspace-relative path.
-
-    The evidence producer should emit workspace-relative paths, but live Codex
-    runs may still report absolute files under the disposable target repo.  Treat
-    those as the same claim only after proving they resolve inside ``task_cwd``.
-    Paths outside the workspace, empty paths, and relative traversal remain
-    unsupported evidence claims.
-    """
-    raw_value = value.strip()
-    if not raw_value or task_cwd is None:
-        return None
-
-    base = Path(task_cwd).resolve()
-    candidate = Path(raw_value)
-    if not candidate.is_absolute() and ".." in candidate.parts:
-        return None
-
-    resolved = (candidate if candidate.is_absolute() else base / candidate).resolve()
-    try:
-        relative = resolved.relative_to(base)
-    except ValueError:
-        return None
-
-    if not relative.parts or ".." in relative.parts:
-        return None
-    return relative.as_posix()
-
-
-def _runtime_support_messages_for_field(
-    field_name: str,
-    messages: tuple[AgentMessage, ...],
-) -> tuple[AgentMessage, ...]:
-    """Narrow support messages for profile-known evidence fields."""
-    normalized = field_name.lower()
-    if normalized == "files_touched":
-        return messages
-    if normalized in {"commands_run", "tests_passed"}:
-        return tuple(message for message in messages if message.tool_name == "Bash")
-    return messages
-
-
-def _runtime_messages_support_claim(value: str, messages: tuple[AgentMessage, ...]) -> bool:
-    """Return True when a non-final runtime message backs a claim string."""
-    needle = value.strip().lower()
-    return bool(needle) and any(
-        needle in _runtime_message_search_text(message) for message in messages
-    )
-
-
-def _runtime_message_supports_command_claim(value: str, message: AgentMessage) -> bool:
-    """Return True when one runtime message backs a command claim.
-
-    Codex commonly records the executed Bash command as a shell wrapper such as
-    ``/bin/zsh -lc 'cd /workspace && python -m unittest "test_hello.py"'``
-    while typed evidence may claim the inner test command.  Treat those as
-    equivalent only through the structured Bash command field; arbitrary output
-    text or assistant narration must not create command aliases.
-    """
-    if message.tool_name != "Bash":
-        return _runtime_messages_support_claim(value, (message,))
-    claim_aliases = set(_normalized_command_claim_aliases(value))
-    claim_test_invocation = _test_command_invocation(value)
-    for runtime_command in _runtime_message_command_values(message):
-        runtime_aliases = set(_normalized_command_claim_aliases(runtime_command))
-        if claim_aliases and runtime_aliases and claim_aliases.intersection(runtime_aliases):
-            return True
-
-        runtime_inner_command = _single_command_after_safe_shell_preamble(runtime_command)
-        if runtime_inner_command and runtime_inner_command in claim_aliases:
-            return True
-
-        runtime_test_invocation = _test_command_invocation(runtime_command)
-        if (
-            claim_test_invocation
-            and runtime_test_invocation
-            and (
-                runtime_test_invocation == claim_test_invocation
-                or runtime_test_invocation.startswith(claim_test_invocation + " ")
-            )
-        ):
-            return True
-    return False
-
-
-def _runtime_messages_support_command_claim(
-    value: str,
-    messages: tuple[AgentMessage, ...],
-) -> bool:
-    """Return True when runtime messages back a command claim."""
-    return any(_runtime_message_supports_command_claim(value, message) for message in messages)
-
-
-def _runtime_messages_have_masked_test_command_form(
-    value: str,
-    messages: tuple[AgentMessage, ...],
-) -> bool:
-    """Return True when a test command claim matches only after unsafe plumbing.
-
-    This deliberately does NOT prove the command claim. It distinguishes a real
-    transcript shape that failed the evidence contract (for example a test run
-    piped through ``tail`` without ``set -o pipefail``) from a fabrication where
-    no related test command appears at all.
-    """
-    claim_invocation = _test_command_invocation(value)
-    if claim_invocation is None:
-        return False
-    for message in messages:
-        if message.tool_name != "Bash":
-            continue
-        for runtime_command in _runtime_message_command_values(message):
-            if not _has_trailing_output_filter_pipeline(runtime_command):
-                continue
-            runtime_invocation = _test_command_invocation_allowing_output_plumbing(runtime_command)
-            if runtime_invocation is None:
-                continue
-            if runtime_invocation == claim_invocation or runtime_invocation.startswith(
-                claim_invocation + " "
-            ):
-                return True
-    return False
-
-
-def _runtime_messages_have_masked_test_command_for_test_claim(
-    *,
-    value: str,
-    messages: tuple[AgentMessage, ...],
-    task_cwd: str | None,
-) -> bool:
-    """Return True when a rejected test claim depends on masked test output.
-
-    This is diagnostic only. It lets the verifier classify a dependent
-    ``tests_passed`` failure with the same evidence-form mismatch as the
-    rejected ``commands_run`` claim, while still refusing to accept the masked
-    command as proof.
-    """
-    for index, message in enumerate(messages):
-        if message.tool_name != "Bash":
-            continue
-        masked_invocations: list[str] = []
-        for runtime_command in _runtime_message_command_values(message):
-            if not _has_trailing_output_filter_pipeline(runtime_command):
-                continue
-            runtime_invocation = _test_command_invocation_allowing_output_plumbing(runtime_command)
-            if runtime_invocation is not None:
-                masked_invocations.append(runtime_invocation)
-        if not masked_invocations:
-            continue
-
-        chunk = [message]
-        for following in messages[index + 1 :]:
-            if following.tool_name and not _is_tool_result_message(following):
-                break
-            chunk.append(following)
-        if not any(_message_contains_test_success(item) for item in chunk):
-            continue
-        chunk_text = "\n".join(_runtime_message_search_text(item) for item in chunk)
-        chunk_test_proof_text = "\n".join(_runtime_message_test_proof_text(item) for item in chunk)
-        if any(
-            _test_command_targets_claim(
-                command=command,
-                claim=value,
-                chunk_text=chunk_text,
-                chunk_test_proof_text=chunk_test_proof_text,
-                messages=messages,
-                task_cwd=task_cwd,
-            )
-            for command in masked_invocations
-        ):
-            return True
-    return False
-
-
-def _runtime_messages_support_file_claim(
-    value: str,
-    messages: tuple[AgentMessage, ...],
-    *,
-    task_cwd: str | None,
-) -> bool:
-    """Return True when runtime transcript evidence backs a workspace file claim.
-
-    Existence alone is not sufficient for ``files_touched``: a stale file in the
-    workspace must not prove that this run created or modified it. Exact
-    transcript support is preferred; basename support is accepted only when the
-    claimed relative path resolves inside the active workspace, which covers
-    tool outputs that report ``generated.py`` instead of ``src/generated.py``.
-    """
-    relative_claim = _workspace_relative_file_claim(value, task_cwd=task_cwd)
-    if relative_claim is None:
-        return False
-    candidate = Path(relative_claim)
-    base = Path(task_cwd).resolve()
-    resolved = (base / candidate).resolve()
-    if any(
-        _runtime_message_supports_file_reference(
-            relative_claim,
-            message,
-            messages=messages,
-            index=index,
-            task_cwd=task_cwd,
-        )
-        for index, message in enumerate(messages)
-    ):
-        return True
-    if not resolved.exists():
-        return False
-    basename = candidate.name.strip().lower()
-    return bool(basename) and any(
-        _runtime_message_supports_file_reference(
-            basename,
-            message,
-            messages=messages,
-            index=index,
-            task_cwd=task_cwd,
-            allow_bash_command_text=False,
-        )
-        for index, message in enumerate(messages)
-    )
-
-
-def _runtime_message_supports_file_reference(
-    reference: str,
-    message: AgentMessage,
-    *,
-    messages: tuple[AgentMessage, ...],
-    index: int,
-    task_cwd: str | None,
-    allow_bash_command_text: bool = True,
-) -> bool:
-    """Return True when one message plausibly reports touching a file reference."""
-    normalized_reference = reference.strip().lower()
-    if not normalized_reference:
-        return False
-    text = _runtime_message_file_proof_text(message)
-    if message.tool_name == "Bash":
-        return _text_supports_file_mutation_reference(text, normalized_reference) or (
-            allow_bash_command_text
-            and _bash_command_mutates_file_reference(message, normalized_reference)
-            and _runtime_message_has_success_evidence(message, messages=messages, index=index)
-        )
-    if message.tool_name in {"Edit", "Write", "NotebookEdit"}:
-        return any(
-            _file_claim_matches_runtime_path(reference, path, task_cwd=task_cwd)
-            for path in _runtime_message_file_path_values(message)
-        )
-    return _text_supports_file_mutation_reference(text, normalized_reference)
-
-
-def _text_supports_file_mutation_reference(text: str, normalized_reference: str) -> bool:
-    """Return True when text pairs a file reference with mutation language."""
-    if not text:
-        return False
-    reference_pattern = _file_reference_pattern(normalized_reference)
-    if not reference_pattern.search(text):
-        return False
-    return bool(
-        re.search(
-            rf"(?<![\w./-]){re.escape(normalized_reference)}(?![\w./-]).*\b("
-            r"updated|modified|changed|created|generated|wrote|written|patched"
-            r")\b|\b("
-            r"updated|modified|changed|created|generated|wrote|written|patched"
-            rf")\b.*(?<![\w./-]){re.escape(normalized_reference)}(?![\w./-])",
-            text,
-        )
-    )
-
-
-def _file_reference_pattern(normalized_reference: str) -> re.Pattern[str]:
-    """Return a conservative token pattern for a workspace-relative file reference."""
-    return re.compile(rf"(?<![\w./-]){re.escape(normalized_reference)}(?![\w./-])")
-
-
-def _bash_command_mutates_file_reference(message: AgentMessage, normalized_reference: str) -> bool:
-    """Return True for explicit shell writes to the referenced file.
-
-    Bash command text is only trusted when the command itself carries mutation
-    semantics for the claimed file. This preserves direct shell-edit evidence
-    such as ``touch src/generated.py`` or ``printf ... > src/generated.py``
-    without allowing read-only probes like ``grep updated src/generated.py`` to
-    prove ``files_touched`` merely by containing a path and a mutation word.
-    """
-    tool_input = message.data.get("tool_input")
-    if not isinstance(tool_input, dict):
-        return False
-    command = tool_input.get("command")
-    if not isinstance(command, str):
-        return False
-    normalized_command = command.strip().lower()
-    if not normalized_command:
-        return False
-    if not _file_reference_pattern(normalized_reference).search(normalized_command):
-        return False
-    quoted_reference = rf"['\"]?{re.escape(normalized_reference)}['\"]?"
-    if re.search(rf"(^|[\s;&|])(?:\d?>|&>|>>|\d>>)\s*{quoted_reference}", normalized_command):
-        return True
-    return bool(
-        re.search(
-            rf"(^|[\s;&|])(touch|truncate|tee)\b[^;&|]*\s{quoted_reference}(?=$|[\s;&|])",
-            normalized_command,
-        )
-        or re.search(
-            rf"(^|[\s;&|])(sed|perl)\b[^;&|]*\s-[^\s;&|]*i[^;&|]*\s"
-            rf"{quoted_reference}(?=$|[\s;&|])",
-            normalized_command,
-        )
-    )
-
-
-def _runtime_message_has_success_signal(message: AgentMessage) -> bool:
-    """Return True when one runtime message carries successful completion evidence."""
-    if message.is_error:
-        return False
-    exit_code = message.data.get("exit_code")
-    if isinstance(exit_code, int):
-        return exit_code == 0
-    if message.data.get("subtype") == "success":
-        return True
-    status = message.data.get("status")
-    if isinstance(status, str) and status.strip().lower() in {
-        "completed",
-        "success",
-        "succeeded",
-    }:
-        return True
-    text = "\n".join(
-        str(part)
-        for part in (
-            message.content,
-            message.data.get("result_preview"),
-            message.data.get("output"),
-            message.data.get("stdout"),
-            message.data.get("stderr"),
-        )
-        if isinstance(part, str)
-    ).lower()
-    return bool(re.search(r"\b(exit\s*code\s*0|completed|succeeded|success)\b", text))
-
-
-def _runtime_message_has_success_evidence(
-    message: AgentMessage,
-    *,
-    messages: tuple[AgentMessage, ...],
-    index: int,
-) -> bool:
-    """Return True when a tool-call message itself or its result proves success."""
-    if _runtime_message_has_success_signal(message):
-        return True
-    return _runtime_message_has_following_success(messages, index)
-
-
-def _runtime_message_has_following_success(messages: tuple[AgentMessage, ...], index: int) -> bool:
-    """Return True when a tool-call message is followed by a successful result."""
-    for candidate in messages[index + 1 :]:
-        if candidate.type == "tool":
-            return False
-        if candidate.is_error:
-            return False
-        if _runtime_message_has_success_signal(candidate):
-            return True
-    return False
-
-
-def _runtime_message_file_proof_text(message: AgentMessage) -> str:
-    """Return text that can prove a file was touched by the current run.
-
-    For Bash tool invocations, command text is not proof by itself: read-only
-    commands such as ``grep updated src/app.py`` can contain both the claimed
-    path and mutation verbs. Trust Bash result/output fields instead. Dedicated
-    edit/write tools still expose their tool inputs because their tool identity
-    supplies the mutation semantics.
-    """
-    if message.tool_name == "Bash":
-        parts: list[str] = []
-        for key in ("result_preview", "output", "stdout", "stderr"):
-            value = message.data.get(key)
-            if isinstance(value, str):
-                parts.append(value)
-        return "\n".join(parts).lower()
-    return _runtime_message_search_text(message)
-
-
-def _looks_like_test_command(command: str) -> bool:
-    """Return True for common whole-suite or targeted test invocations."""
-    return _test_command_invocation(command) is not None
-
-
-def _test_command_invocation(command: str) -> str | None:
-    """Return the backed inner test invocation for a direct or wrapped command.
-
-    Output plumbing (``2>&1``, ``| tail -20``) is peeled first so a clean
-    invocation can be extracted from a ``<cmd> 2>&1 | tail -20`` runtime
-    command. ``_strip_command_output_plumbing`` is deliberately narrow — only
-    presentation-only tails are peeled — so an evidence-altering filter such
-    as ``| grep passed`` survives the strip and is rejected downstream by
-    ``_test_invocation_from_prefix`` rather than being silently dropped.
-    """
-    normalized = command.strip().lower()
-    if not normalized:
-        return None
-
-    direct_candidate = _strip_command_output_plumbing(normalized)
-    if (
-        _has_trailing_output_filter_pipeline(normalized)
-        and not _output_filter_pipeline_is_pipefail_protected(normalized)
-        and _test_invocation_from_prefix(direct_candidate) is not None
-    ):
-        direct_candidate = normalized
-    direct = _test_invocation_from_prefix(direct_candidate)
-    if direct is not None:
-        return direct
-
-    body = _shell_command_body(normalized)
-    if body is None:
-        return None
-    return _test_invocation_from_shell_body(body)
-
-
-def _test_command_invocation_allowing_output_plumbing(command: str) -> str | None:
-    """Return a test invocation after stripping output plumbing unconditionally.
-
-    This must not be used as command proof. It exists only to classify rejected
-    evidence forms for diagnostics while preserving the #1208 masking guard.
-    """
-    normalized = command.strip().lower()
-    if not normalized:
-        return None
-    direct = _test_invocation_from_prefix(_strip_command_output_plumbing(normalized))
-    if direct is not None:
-        return direct
-    body = _shell_command_body(normalized)
-    if body is None:
-        return None
-    for segment, _pipefail_enabled in _segments_after_safe_shell_preamble_with_pipefail(body):
-        invocation = _test_invocation_from_prefix(_strip_command_output_plumbing(segment))
-        if invocation is not None:
-            return invocation
-        return None
-    return None
-
-
-def _shell_command_body(command: str) -> str | None:
-    """Return the ``-c`` body when the command starts with a shell wrapper."""
-    try:
-        parts = shlex.split(command)
-    except ValueError:
-        return None
-    if len(parts) < 3:
-        return None
-    shell_name = Path(parts[0]).name
-    if shell_name not in {"bash", "zsh", "sh"}:
-        return None
-    option_index = next(
-        (index for index, part in enumerate(parts[1:], start=1) if part in {"-c", "-lc", "-cl"}),
-        None,
-    )
-    if option_index is None or option_index + 1 >= len(parts):
-        return None
-    return parts[option_index + 1].strip()
-
-
-def _test_invocation_from_shell_body(body: str) -> str | None:
-    """Return a test invocation after conservative shell setup preambles."""
-    for segment, pipefail_enabled in _segments_after_safe_shell_preamble_with_pipefail(body):
-        candidate = _strip_command_output_plumbing(segment)
-        if (
-            _has_trailing_output_filter_pipeline(segment)
-            and not pipefail_enabled
-            and _test_invocation_from_prefix(candidate) is not None
-        ):
-            candidate = segment
-        invocation = _test_invocation_from_prefix(candidate)
-        if invocation is not None:
-            return invocation
-        return None
-    return None
-
-
-def _single_command_after_safe_shell_preamble(command: str) -> str | None:
-    """Return a wrapped inner command after only safe setup preambles.
-
-    Generic ``commands_run`` evidence may cite the useful command inside a
-    runtime-recorded shell wrapper such as ``cd /work && python scripts/gen.py``.
-    Keep this narrower than substring containment: only ignore setup-only
-    preambles and only when exactly one non-preamble command remains.
-    """
-    body = _shell_command_body(command)
-    if body is None:
-        return None
-    segments = tuple(_segments_after_safe_shell_preamble(body))
-    if len(segments) != 1:
-        return None
-    segment = segments[0]
-    stripped = _strip_command_output_plumbing(segment)
-    if (
-        _has_trailing_output_filter_pipeline(segment)
-        and not _output_filter_pipeline_is_pipefail_protected(body)
-        and _looks_like_test_command(stripped)
-    ):
-        return None
-    return _normalized_evidence_text(stripped)
-
-
-def _segments_after_safe_shell_preamble(body: str) -> tuple[str, ...]:
-    """Return non-preamble shell segments after setup-only commands."""
-    return tuple(
-        segment
-        for segment, _pipefail_enabled in _segments_after_safe_shell_preamble_with_pipefail(body)
-    )
-
-
-def _segments_after_safe_shell_preamble_with_pipefail(body: str) -> tuple[tuple[str, bool], ...]:
-    """Return non-preamble segments with pipefail state active before each one."""
-    remaining: list[tuple[str, bool]] = []
-    pipefail_enabled = False
-    for segment in re.split(r"\s*&&\s*", body.strip()):
-        normalized_segment = segment.strip()
-        if not normalized_segment:
-            continue
-        if not remaining and _is_safe_test_command_preamble(normalized_segment):
-            if _is_pipefail_preamble(normalized_segment):
-                pipefail_enabled = True
-            continue
-        remaining.append((normalized_segment, pipefail_enabled))
-    return tuple(remaining)
-
-
-def _is_safe_test_command_preamble(segment: str) -> bool:
-    """Return True for shell setup segments that do not execute tests themselves."""
-    try:
-        parts = shlex.split(segment)
-    except ValueError:
-        return False
-    if not parts:
-        return True
-    if parts[0] == "cd" and len(parts) == 2:
-        return True
-    if _is_pipefail_parts(parts):
-        return True
-    if parts[0] == "export" and len(parts) > 1:
-        return all(_is_env_assignment(part) for part in parts[1:])
-    return all(_is_env_assignment(part) for part in parts)
-
-
-def _is_env_assignment(value: str) -> bool:
-    """Return True for a simple shell environment assignment token."""
-    return bool(re.match(r"^[A-Za-z_][A-Za-z0-9_]*=.*$", value))
-
-
-def _strip_env_prefix(parts: list[str]) -> list[str]:
-    """Remove leading env assignment tokens before command recognition."""
-    index = 0
-    if parts and parts[0] == "env":
-        index = 1
-    while index < len(parts) and _is_env_assignment(parts[index]):
-        index += 1
-    return parts[index:]
-
-
-def _has_gradle_or_maven_test_skip(parts: list[str]) -> bool:
-    """Return True when a Gradle/Maven command explicitly disables tests."""
-
-    def maven_skip_property_disables_tests(value: str) -> bool:
-        normalized_value = value.lower()
-        if normalized_value in {"skiptests", "maven.test.skip"}:
-            return True
-        if normalized_value.startswith("skiptests=") or normalized_value.startswith(
-            "maven.test.skip="
-        ):
-            _, _, property_value = normalized_value.partition("=")
-            return property_value not in {"false", "0", "no", "off"}
-        return False
-
-    for index, part in enumerate(parts):
-        normalized = part.lower()
-        if normalized == "-d" and index + 1 < len(parts):
-            if maven_skip_property_disables_tests(parts[index + 1]):
-                return True
-        if normalized == "--define" and index + 1 < len(parts):
-            if maven_skip_property_disables_tests(parts[index + 1]):
-                return True
-        if normalized.startswith("--define="):
-            _, _, define_value = normalized.partition("=")
-            if maven_skip_property_disables_tests(define_value):
-                return True
-        if normalized.startswith("-d") and maven_skip_property_disables_tests(normalized[2:]):
-            return True
-        if normalized == "--exclude-task" and index + 1 < len(parts):
-            excluded_task = parts[index + 1].lower().lstrip(":")
-            if excluded_task == "test" or excluded_task.endswith(":test"):
-                return True
-        if normalized.startswith("--exclude-task="):
-            _, _, excluded_task = normalized.partition("=")
-            excluded_task = excluded_task.lstrip(":")
-            if excluded_task == "test" or excluded_task.endswith(":test"):
-                return True
-        if normalized == "-x" and index + 1 < len(parts):
-            excluded_task = parts[index + 1].lower().lstrip(":")
-            if excluded_task == "test" or excluded_task.endswith(":test"):
-                return True
-        if normalized.startswith("-x") and len(normalized) > 2:
-            excluded_task = normalized[2:].lstrip(":")
-            if excluded_task == "test" or excluded_task.endswith(":test"):
-                return True
-    return False
-
-
-def _test_invocation_from_prefix(command: str) -> str | None:
-    """Return a normalized test invocation only when it starts the command text.
-
-    Refuses to extract from commands that still contain a residual shell pipe
-    after presentation plumbing has been peeled (``pytest x | grep passed``).
-    A residual pipe means the runtime command is followed by an
-    evidence-transforming filter (``grep`` / ``wc`` / ``tee``); treating the
-    bare prefix as the clean test invocation there would let a filtered run
-    silently back a clean ``tests_passed`` / ``commands_run`` claim via the
-    ``startswith`` widening in ``_runtime_message_supports_command_claim``.
-    """
-    try:
-        parts = shlex.split(command)
-    except ValueError:
-        parts = command.replace('"', "").replace("'", "").split()
-    parts = _strip_env_prefix(parts)
-    if not parts:
-        return None
-    if "|" in parts:
-        return None
-
-    if parts[0] in {"pytest", "py.test", "tox", "nox"}:
-        return _normalized_evidence_text(" ".join(parts))
-    if len(parts) >= 2 and parts[0] in {"npm", "pnpm", "yarn"} and parts[1] == "test":
-        return _normalized_evidence_text(" ".join(parts))
-    if len(parts) >= 3 and parts[:3] == ["uv", "run", "pytest"]:
-        return _normalized_evidence_text(" ".join(parts))
-    if (
-        len(parts) >= 3
-        and parts[:2] == ["python", "-m"]
-        and parts[2]
-        in {
-            "pytest",
-            "unittest",
-        }
-    ):
-        return _normalized_evidence_text(" ".join(parts))
-    executable = Path(parts[0]).name
-    if (
-        executable in {"gradle", "gradlew", "mvn", "mvnw"}
-        and not _has_gradle_or_maven_test_skip(parts[1:])
-        and any(part in {"test", "check", "verify"} or part.endswith(":test") for part in parts[1:])
-    ):
-        return _normalized_evidence_text(" ".join(parts))
-    return None
-
-
-def _unittest_command_invocation(command: str) -> str | None:
-    """Return the embedded ``python -m unittest`` invocation, if present."""
-    invocation = _test_command_invocation(command)
-    if invocation is None:
-        return None
-    parts = invocation.split()
-    if len(parts) >= 3 and parts[:3] == ["python", "-m", "unittest"]:
-        return invocation
-    return None
-
-
-def _looks_like_unittest_command(command: str) -> bool:
-    """Return True when a shell command invokes stdlib unittest."""
-    return _unittest_command_invocation(command) is not None
-
-
-# Output-only shell filters: a trailing pipe into one of these is presentation
-# or paging, not the work an evidence claim is about.
-#
-# Deliberately narrow: only filters that pass the output stream through (or
-# truncate it positionally) are allowed. ``grep``/``egrep``/``fgrep`` are
-# excluded because they can hide failure lines and make a filtered run back a
-# clean ``commands_run`` / ``tests_passed`` claim (e.g. ``pytest ... | grep
-# passed``). ``tee`` and ``wc`` are excluded for the same reason: ``tee`` can
-# redirect the stream and ``wc`` collapses it to a count, both of which alter
-# what the runtime would have observed and so weaken anti-fabrication.
-_OUTPUT_FILTER_COMMANDS = frozenset({"tail", "head", "cat", "less", "more"})
-
-# Trailing shell output redirection (``2>&1``, ``> log``, ``2> err``, ``&> out``).
-_TRAILING_REDIRECT_RE = re.compile(
-    r"\s*(?:[0-9]*>{1,2}\s*(?:&[0-9]+|[^\s|]+)|&>{1,2}\s*[^\s|]+)\s*$"
-)
-
-
-def _normalized_shell_words_text(command: str) -> str | None:
-    """Return a quote-insensitive normalized argv spelling for one shell command.
-
-    This keeps command evidence matching exact at the argv level while allowing
-    common shell spelling differences such as ``--tests "ClassName"`` versus
-    ``--tests ClassName``. Commands that cannot be parsed, still contain a
-    pipeline, or contain shell control operators are left to the stricter raw
-    aliases.
-    """
-    text = command.strip()
-    if not text:
-        return None
-    try:
-        parts = shlex.split(text)
-    except ValueError:
-        return None
-    if not parts or any(part in {"|", "&&", ";", "||"} for part in parts):
-        return None
-    return _normalized_evidence_text(" ".join(parts))
-
-
-def _strip_command_output_plumbing(command: str) -> str:
-    """Return a command with trailing output redirection and pager pipes removed.
-
-    Agents routinely run ``<cmd> 2>&1 | tail -20`` while their ``commands_run``
-    evidence cites the clean ``<cmd>``. The trailing redirection and the
-    output-only pager pipe are presentation plumbing, not the work being
-    claimed, so they must not block a match. Deliberately conservative:
-
-    - Only trailing output redirections (``2>&1``, ``> log``, ``2> err``,
-      ``&> out``) and pipes into a pager-style filter listed in
-      ``_OUTPUT_FILTER_COMMANDS`` (``tail``/``head``/``cat``/``less``/``more``)
-      are dropped. These pass the underlying stream through (or truncate it
-      positionally), so the runtime evidence is unchanged in kind.
-    - Filters that *transform* the stream — ``grep`` family, ``wc``, ``tee`` —
-      are intentionally not stripped, because they can hide failure lines,
-      collapse the stream to a count, or divert it to a file, which would let
-      a filtered run back a clean ``commands_run`` / ``tests_passed`` claim.
-    - Meaningful pipelines such as ``a | python process.py`` are kept, so a
-      partial ``a`` claim is still not proven by an ``a | python process.py``
-      runtime command.
-    """
-    text = command.strip()
-    if not text:
-        return text
-    # Peel output-only filter pipes from the tail (``... | tail -n 20``).
-    while "|" in text:
-        head, _, tail_segment = text.rpartition("|")
-        tail_tokens = tail_segment.split()
-        if tail_tokens and tail_tokens[0].lower() in _OUTPUT_FILTER_COMMANDS:
-            text = head.strip()
-            continue
-        break
-    # Peel trailing output redirections, possibly several (``2>&1 > log``).
-    prev: str | None = None
-    while prev != text:
-        prev = text
-        text = _TRAILING_REDIRECT_RE.sub("", text).strip()
-    return text
-
-
-def _has_trailing_output_filter_pipeline(command: str) -> bool:
-    """Return True when ``command`` ends in a pager-style output pipe."""
-    text = command.strip()
-    while "|" in text:
-        head, _, tail_segment = text.rpartition("|")
-        tail_tokens = tail_segment.split()
-        if tail_tokens and tail_tokens[0].lower() in _OUTPUT_FILTER_COMMANDS:
-            return True
-        text = head.strip()
-    return False
-
-
-def _output_filter_pipeline_is_pipefail_protected(command: str) -> bool:
-    """Return True when pipefail is enabled before the first stripped pipeline."""
-    pipefail_enabled = False
-    for segment in re.split(r"\s*(?:&&|;)\s*", command.strip()):
-        normalized_segment = segment.strip()
-        if not normalized_segment:
-            continue
-        if _is_pipefail_preamble(normalized_segment):
-            pipefail_enabled = True
-            continue
-        if _has_trailing_output_filter_pipeline(normalized_segment):
-            return pipefail_enabled
-    return False
-
-
-def _uses_pipefail(command: str) -> bool:
-    """Return True when shell text explicitly preserves upstream pipe status."""
-    for segment in re.split(r"\s*(?:&&|;)\s*", command.strip()):
-        try:
-            parts = shlex.split(segment)
-        except ValueError:
-            continue
-        if _is_pipefail_parts(parts):
-            return True
-    return False
-
-
-def _is_pipefail_preamble(segment: str) -> bool:
-    try:
-        parts = shlex.split(segment)
-    except ValueError:
-        return False
-    return _is_pipefail_parts(parts)
-
-
-def _is_pipefail_parts(parts: list[str]) -> bool:
-    return parts == ["set", "-o", "pipefail"]
-
-
-def _normalized_command_claim_aliases(command: str) -> tuple[str, ...]:
-    """Return normalized command forms that a concise evidence claim may use.
-
-    Structured Bash tool inputs may wrap the user command as
-    ``/bin/zsh -lc '<body>'``.  The wrapper itself is runtime-backed, so an
-    evidence claim may cite the exact shell body without re-stating the wrapper.
-    Keep this alias exact: test-command-specific helpers handle conservative
-    setup preambles, while generic ``commands_run`` claims should not be proven
-    by partial substrings of arbitrary shell scripts.
-    """
-    normalized = _normalized_evidence_text(command)
-    aliases = [normalized] if normalized else []
-
-    def append_alias(candidate: str | None) -> None:
-        if candidate and candidate not in aliases:
-            aliases.append(candidate)
-
-    append_alias(_normalized_shell_words_text(command))
-    shell_body = _shell_command_body(command)
-    normalized_shell_body = _normalized_evidence_text(shell_body) if shell_body else None
-    append_alias(normalized_shell_body)
-    append_alias(_normalized_shell_words_text(shell_body) if shell_body else None)
-    test_invocation = _test_command_invocation(command)
-    append_alias(test_invocation)
-    # A recorded command may append output plumbing (``... 2>&1 | tail -20``)
-    # that a concise ``commands_run`` claim omits. Add plumbing-stripped variants
-    # so the two still match. Alias matching stays exact (set intersection), so
-    # this does not widen proof to arbitrary substrings. Also add argv-normalized
-    # plumbing-stripped forms so quoted arguments in the runtime command match
-    # unquoted evidence claims for the same argv vector.
-    for base in tuple(aliases):
-        stripped_raw = _strip_command_output_plumbing(base)
-        stripped = _normalized_evidence_text(stripped_raw)
-        if (
-            stripped
-            and stripped != base
-            and _has_trailing_output_filter_pipeline(base)
-            and _looks_like_test_command(stripped)
-            and not _output_filter_pipeline_is_pipefail_protected(base)
-        ):
-            continue
-        append_alias(stripped)
-        append_alias(_normalized_shell_words_text(stripped_raw))
-    return tuple(aliases)
-
-
-def _text_contains_unittest_success(text: str) -> bool:
-    """Return True for real unittest success output."""
-    return _text_contains_test_success(text) and bool(
-        re.search(r"\bran\s+[1-9]\d*\s+tests?\b[\s\S]*\bok\b", text.lower())
-    )
-
-
-def _text_contains_test_success(text: str) -> bool:
-    """Return True when text contains a conservative test-success signal."""
-    text = text.lower()
-    zero_failure_pattern = (
-        r"\b(0\s+(failed|failures?|errors?)|"
-        r"(failed|failures?|errors?)\s*[:=]\s*0|"
-        r"no\s+(tests?\s+)?(failed|failures?|errors?))\b"
-    )
-    failure_scan_text = re.sub(zero_failure_pattern, "", text)
-    if re.search(
-        r"\b[1-9]\d*\s+(failed|failures?|errors?)\b|"
-        r"\b(failed|failure|failures?|error|errors)\b|"
-        r"exit\s*code\s*[1-9]",
-        failure_scan_text,
-    ):
-        return False
-    if re.search(r"\b0\s+passed\b", text) and not re.search(r"\b[1-9]\d*\s+passed\b", text):
-        return False
-    if re.search(r"\btask\s+[:\w.-]*test\b[^\n]*(no-source|skipped)\b", text):
-        return False
-    if re.search(r"\b0\s+tests?\s+(completed|run|executed)\b", text):
-        return False
-    if re.search(r"\bno\s+tests?\s+(found|run|executed)\b", text):
-        return False
-    return bool(
-        re.search(
-            r"\b([1-9]\d*\s+passed|passed|pass|success|successful|succeeded)\b|"
-            r"\bbuild\s+successful\b|exit\s*code\s*0",
-            text,
-        )
-        or re.search(r"\bran\s+[1-9]\d*\s+tests?\b[\s\S]*\bok\b", text)
-    )
-
-
-def _message_contains_test_success(message: AgentMessage) -> bool:
-    """Return True when one message says a test command passed."""
-    parts = [message.content]
-    for key in ("result_preview", "output", "stdout", "status", "subtype"):
-        value = message.data.get(key)
-        if isinstance(value, str):
-            parts.append(value)
-    exit_code = message.data.get("exit_code")
-    if type(exit_code) is int:
-        parts.append(f"exit code {exit_code}")
-    return _text_contains_test_success("\n".join(parts))
-
-
-def _runtime_message_test_proof_text(message: AgentMessage) -> str:
-    """Return runtime-produced text that can prove test output for a Bash chunk.
-
-    Assistant narration after a Bash call is useful transcript context, but it
-    is not runtime output for that command. Keep summary matching tied to the
-    Bash output/result payloads and tool-result messages that runtimes emit.
-    """
-    resultish = (
-        message.type in {"result", "tool_result"} or message.data.get("subtype") == "tool_result"
-    )
-    parts: list[str] = []
-    if resultish:
-        parts.append(message.content)
-    for key in ("result_preview", "output", "stdout", "stderr", "tool_result_text"):
-        value = message.data.get(key)
-        if isinstance(value, str):
-            parts.append(value)
-    tool_result = message.data.get("tool_result")
-    if isinstance(tool_result, dict):
-        for key in ("text_content", "content", "output", "stdout", "stderr"):
-            value = tool_result.get(key)
-            if isinstance(value, str):
-                parts.append(value)
-    elif isinstance(tool_result, str):
-        parts.append(tool_result)
-    return "\n".join(parts)
-
-
-def _is_tool_result_message(message: AgentMessage) -> bool:
-    """Return True for runtime tool-result messages, including named-tool variants."""
-    return message.type in {"result", "tool_result"} or message.data.get("subtype") == "tool_result"
-
-
-def _test_claim_file_part(value: str) -> str | None:
-    """Return the file path portion of a pytest node-id style claim."""
-    stripped = value.strip()
-    if not stripped:
-        return None
-    file_part = stripped.split("::", 1)[0].strip()
-    return file_part or None
-
-
-def _normalized_evidence_text(text: str) -> str:
-    """Normalize transcript/claim text for conservative containment checks."""
-    return " ".join(text.lower().split())
-
-
-def _claim_summary_matches_runtime_chunk(
-    *,
-    command: str,
-    claim: str,
-    chunk_text: str,
-) -> bool:
-    """Return True when a command+summary claim is present in runtime output.
-
-    This keeps the verifier transcript-driven: the claim may combine the backed
-    command and a unittest-style success summary, but the summary itself must
-    also appear in the runtime chunk. The claim text alone is never proof.
-    """
-    normalized_claim = _normalized_evidence_text(claim)
-    normalized_chunk = _normalized_evidence_text(chunk_text)
-    summary = ""
-    for normalized_command in _normalized_command_claim_aliases(command):
-        if normalized_command in normalized_claim:
-            summary = normalized_claim.split(normalized_command, 1)[1].strip(" :-")
-            break
-    if not summary or summary not in normalized_chunk:
-        return False
-    if (
-        summary == "ok"
-        and _looks_like_unittest_command(command)
-        and _text_contains_unittest_success(chunk_text)
-    ):
-        return True
-    return _text_contains_test_success(summary)
-
-
-def _claim_contains_command_success_summary(*, command: str, claim: str) -> bool:
-    """Return True when a test claim appends a success summary to a command."""
-    normalized_claim = _normalized_evidence_text(claim)
-    for normalized_command in _normalized_command_claim_aliases(command):
-        if normalized_command in normalized_claim:
-            summary = normalized_claim.split(normalized_command, 1)[1].strip(" :-")
-            return bool(summary) and _text_contains_test_success(summary)
-    return False
-
-
-def _test_command_targets_claim(
-    *,
-    command: str,
-    claim: str,
-    chunk_text: str,
-    chunk_test_proof_text: str,
-    messages: tuple[AgentMessage, ...],
-    task_cwd: str | None,
-) -> bool:
-    """Return True when a successful test command can cover a test claim."""
-    needle = claim.strip().lower()
-    if _claim_contains_command_success_summary(command=command, claim=claim):
-        return _claim_summary_matches_runtime_chunk(
-            command=command,
-            claim=claim,
-            chunk_text=chunk_test_proof_text,
-        )
-    if needle and needle in chunk_text:
-        return True
-
-    file_part = _test_claim_file_part(claim)
-    if file_part is None:
-        return False
-    normalized_file = file_part.lower()
-    normalized_command = command.lower()
-    if normalized_file in chunk_text or normalized_file in normalized_command:
-        return True
-    if _claim_summary_matches_runtime_chunk(
-        command=command,
-        claim=claim,
-        chunk_text=chunk_test_proof_text,
-    ):
-        return True
-
-    # A broad suite command such as ``pytest`` can cover a node-id claim when
-    # the claimed test file is also backed by current-run mutation evidence.
-    # Existence alone is deliberately insufficient: otherwise a transcript with
-    # unrelated ``pytest`` output could prove any stale test file in the tree.
-    command_parts = (_test_command_invocation(command) or normalized_command).split()
-    broad_pytest = command_parts in (["pytest"], ["py.test"]) or command_parts[-3:] == [
-        "python",
-        "-m",
-        "pytest",
-    ]
-    if not broad_pytest or task_cwd is None:
-        return False
-    return _runtime_messages_support_file_claim(file_part, messages, task_cwd=task_cwd)
-
-
-def _runtime_messages_support_test_claim(
-    *,
-    value: str,
-    backed_commands: tuple[str, ...],
-    messages: tuple[AgentMessage, ...],
-    task_cwd: str | None,
-) -> bool:
-    """Return True when a backed test command chunk proves one test claim."""
-    needle = value.strip().lower()
-    if not needle:
-        return False
-    for index, message in enumerate(messages):
-        if message.tool_name != "Bash":
-            continue
-        # Candidate test commands are drawn from two transcript-grounded
-        # sources: (1) ``commands_run`` evidence entries already proven against
-        # the transcript, and (2) the Bash message's own recorded command. The
-        # latter is backed by definition — it is the literal invocation in the
-        # transcript — so a real ``pytest <file>`` run can support a node-id
-        # ``tests_passed`` claim even when the agent did not also echo that
-        # exact command into its ``commands_run`` evidence.
-        #
-        # These are NOT three independent checks. For a per-message candidate
-        # the ``_runtime_message_supports_command_claim`` gate below is
-        # tautological — the candidate is that message's own command, so it
-        # trivially supports itself. The anti-fabrication guarantee is carried
-        # entirely by the downstream gates: ``_message_contains_test_success``
-        # (reads only structured runtime output, never agent narration) and
-        # ``_test_command_targets_claim`` (anchors the claim's node-id/file to
-        # the recorded command + proof text). The ``_looks_like_test_command``
-        # filter still excludes non-test commands from this candidate source.
-        candidate_commands = (*backed_commands, *_runtime_message_command_values(message))
-        matching_commands = tuple(
-            candidate
-            for candidate in candidate_commands
-            if _looks_like_test_command(candidate)
-            and _runtime_message_supports_command_claim(candidate, message)
-        )
-        if not matching_commands:
-            continue
-        chunk = [message]
-        for following in messages[index + 1 :]:
-            if following.tool_name and not _is_tool_result_message(following):
-                break
-            chunk.append(following)
-        if not any(_message_contains_test_success(item) for item in chunk):
-            continue
-        chunk_text = "\n".join(_runtime_message_search_text(item) for item in chunk)
-        chunk_test_proof_text = "\n".join(_runtime_message_test_proof_text(item) for item in chunk)
-        if any(
-            _test_command_targets_claim(
-                command=command,
-                claim=value,
-                chunk_text=chunk_text,
-                chunk_test_proof_text=chunk_test_proof_text,
-                messages=messages,
-                task_cwd=task_cwd,
-            )
-            for command in matching_commands
-        ):
-            return True
-    return False
-
-
-_REUSABLE_RUNTIME_EVENT_TYPES = frozenset(
-    {
-        "execution.session.recovered",
-        "execution.session.started",
-        "execution.session.resumed",
-    }
-)
-_NON_REUSABLE_RUNTIME_EVENT_TYPES = frozenset(
-    {
-        "execution.session.completed",
-        "execution.session.failed",
-    }
-)
-_AC_RUNTIME_OWNERSHIP_METADATA_KEYS = frozenset(
-    {
-        "ac_id",
-        "ac_index",
-        "attempt_number",
-        "depth",
-        "display_path",
-        "execution_id",
-        "identity_model",
-        "legacy_node_id",
-        "legacy_node_aliases",
-        "legacy_parent_node_id",
-        "legacy_parent_node_aliases",
-        "legacy_session_scope_id",
-        "legacy_session_scope_ids",
-        "legacy_session_state_path",
-        "legacy_session_state_paths",
-        "node_kind",
-        "node_id",
-        "ordinal",
-        "parent_ac_index",
-        "parent_node_id",
-        "path",
-        "retry_attempt",
-        "root_ac_index",
-        "root_ac_number",
-        "scope",
-        "schema_version",
-        "session_attempt_id",
-        "session_role",
-        "session_scope_id",
-        "session_state_path",
-        "sub_ac_index",
-    }
-)
-_AC_RUNTIME_SCOPE_METADATA_KEYS = frozenset(
-    {
-        "ac_id",
-        "ac_index",
-        "execution_id",
-        "identity_model",
-        "node_kind",
-        "node_id",
-        "parent_ac_index",
-        "parent_node_id",
-        "path",
-        "root_ac_index",
-        "scope",
-        "schema_version",
-        "session_role",
-        "session_scope_id",
-        "session_state_path",
-        "sub_ac_index",
-    }
-)
-_AC_RUNTIME_RESUME_METADATA_KEYS = frozenset({"runtime_event_type", "server_session_id"})
-
-# Stall detection constants
-STALL_TIMEOUT_SECONDS: float = 900.0  # 15 minutes of silence → stall for realistic test suites
-HEARTBEAT_INTERVAL_SECONDS: float = 30.0  # Heartbeat emission interval
-MAX_STALL_RETRIES: int = 2  # Max retries after stall (3 total attempts)
-_STALL_SENTINEL = "__STALL_DETECTED__"  # Sentinel error for stall results
-
-# Memory-pressure gate constants
-_MIN_FREE_MEMORY_GB = 2.0
-_MEMORY_CHECK_INTERVAL_SECONDS = 5.0
-_MEMORY_WAIT_MAX_SECONDS = 120.0
-_MAX_LEAF_RESULT_CHARS = 1200
-_SIBLING_HEADLINE_CHARS = 80
-_SiblingACRef = tuple[int | None, str]
-
-
-def _build_governed_parent_summary(level_contexts: list[LevelContext] | None) -> str:
-    """Render level context for governed dispatch without nested H2 headings.
-
-    ``build_context_prompt()`` is also used by the legacy prompt path, where its
-    top-level ``##`` sections are appropriate.  Governed dispatch already wraps
-    that text in ``## Parent context`` via ``compose_context()``, so preserving
-    those headings creates a hard-to-scan nested hierarchy.  Build the governed
-    variant from structured ``LevelContext`` data so only orchestrator-owned
-    section wrappers become compact labels; embedded markdown from AC outputs
-    remains byte-for-byte content inside its original summary text.
-    """
-    if not level_contexts:
-        return ""
-
-    sections: list[str] = []
-    for ctx in level_contexts:
-        text = ctx.to_prompt_text()
-        if text:
-            sections.append(text)
-
-    has_reviews = any(ctx.coordinator_review for ctx in level_contexts)
-    if not sections and not has_reviews:
-        return ""
-
-    lines: list[str] = []
-    if sections:
-        lines.extend(
-            (
-                "Previous Work Context:",
-                "The following ACs have already been completed. "
-                "Use this context to inform your work.",
-                "",
-                "\n\n".join(sections),
-            )
-        )
-
-    for ctx in level_contexts:
-        if ctx.coordinator_review:
-            review = ctx.coordinator_review
-            review_lines: list[str] = []
-
-            if review.review_summary:
-                review_lines.append(f"**Review**: {review.review_summary}")
-
-            if review.fixes_applied:
-                fixes = "; ".join(review.fixes_applied)
-                review_lines.append(f"**Fixes applied**: {fixes}")
-
-            if review.warnings_for_next_level:
-                for warning in review.warnings_for_next_level:
-                    review_lines.append(f"- WARNING: {warning}")
-
-            if review_lines:
-                if lines:
-                    lines.append("")
-                lines.append(f"Coordinator Review (Level {review.level_number}):")
-                lines.extend(review_lines)
-
-    return "\n".join(lines).strip()
-
-
-def _get_available_memory_gb() -> float | None:
-    """Get available memory in GB. Returns None if check fails."""
-    system = platform.system()
-    try:
-        if system == "Darwin":
-            result = subprocess.run(
-                ["vm_stat"],
-                capture_output=True,
-                text=True,
-                timeout=5,
-            )
-            if result.returncode != 0:
-                return None
-            pages_free = 0
-            pages_inactive = 0
-            page_size = 4096  # macOS default
-            for line in result.stdout.splitlines():
-                if "page size of" in line:
-                    parts = line.split()
-                    for part in parts:
-                        if part.isdigit():
-                            page_size = int(part)
-                elif line.startswith("Pages free:"):
-                    pages_free = int(line.split(":")[1].strip().rstrip("."))
-                elif line.startswith("Pages inactive:"):
-                    pages_inactive = int(line.split(":")[1].strip().rstrip("."))
-            return (pages_free + pages_inactive) * page_size / (1024**3)
-
-        elif system == "Linux":
-            with open("/proc/meminfo") as f:
-                for line in f:
-                    if line.startswith("MemAvailable:"):
-                        kb = int(line.split()[1])
-                        return kb / (1024**2)
-            return None
-
-        else:
-            return None
-    except (OSError, ValueError, subprocess.TimeoutExpired):
-        return None
-
-
-# Data models re-exported from parallel_executor_models for backwards compat.
-# (ACExecutionOutcome, ACExecutionResult, ParallelExecutionStageResult,
-#  ParallelExecutionResult, StageExecutionOutcome are imported above.)
-
-
-def _normalize_command(command: str) -> str:
-    """Normalize Bash commands for stable audit output."""
-    return " ".join(command.split())
-
-
-def _normalize_exact_command(command: str) -> str:
-    """Normalize command whitespace while preserving case-sensitive exactness."""
-    return " ".join(command.split())
-
-
-def _truncate_text(text: str, limit: int = _MAX_LEAF_RESULT_CHARS) -> str:
-    """Truncate long evidence blocks while preserving their beginning."""
-    stripped = text.strip()
-    if len(stripped) <= limit:
-        return stripped
-    return stripped[:limit].rstrip() + "\n[TRUNCATED]"
-
-
-def _extract_leaf_evidence_lines(result: ACExecutionResult) -> list[str]:
-    """Extract normalized command, file, and result evidence for a leaf AC."""
-    lines: list[str] = []
-    seen_commands: set[str] = set()
-    seen_file_ops: set[tuple[str, str]] = set()
-
-    for message in result.messages:
-        if not message.tool_name:
-            continue
-        tool_input = message.data.get("tool_input", {})
-        if not isinstance(tool_input, dict):
-            tool_input = {}
-
-        if message.tool_name == "Bash":
-            command = tool_input.get("command")
-            if isinstance(command, str):
-                normalized = _normalize_command(command)
-                if normalized and normalized not in seen_commands:
-                    if not lines:
-                        lines.append("Commands Run:")
-                    lines.append(f"- Bash: {normalized}")
-                    seen_commands.add(normalized)
-            continue
-
-        if message.tool_name in ("Write", "Edit", "NotebookEdit"):
-            path_key = "notebook_path" if message.tool_name == "NotebookEdit" else "file_path"
-            file_path = tool_input.get(path_key)
-            if isinstance(file_path, str) and file_path:
-                file_op = (message.tool_name, file_path)
-                if file_op not in seen_file_ops:
-                    if "File Changes:" not in lines:
-                        lines.append("File Changes:")
-                    lines.append(f"- {message.tool_name}: {file_path}")
-                    seen_file_ops.add(file_op)
-
-    result_text = result.final_message or (f"Error: {result.error}" if result.error else "")
-    if result_text:
-        lines.append("Result:")
-        lines.append(_truncate_text(result_text))
-    return lines
-
-
-def _successful_runtime_test_commands(messages: tuple[AgentMessage, ...]) -> set[str]:
-    """Return Bash test commands backed by adjacent runtime success output."""
-    commands: set[str] = set()
-    for index, message in enumerate(messages):
-        if message.tool_name != "Bash":
-            continue
-        message_commands = {
-            alias
-            for command in _runtime_message_command_values(message)
-            if _looks_like_test_command(command)
-            for alias in _runtime_command_evidence_aliases(command)
-        }
-        if not message_commands:
-            continue
-        chunk = [message]
-        for following in messages[index + 1 :]:
-            if following.tool_name and not _is_tool_result_message(following):
-                break
-            chunk.append(following)
-        if any(
-            not item.is_final
-            and _text_contains_test_success(_runtime_message_test_proof_text(item))
-            for item in chunk
-        ):
-            commands.update(command for command in message_commands if command)
-    return commands
-
-
-def _add_runtime_command_evidence(commands: set[str], command: str) -> None:
-    """Add runtime command evidence without accepting compound shell aliases."""
-    commands.update(_runtime_command_evidence_aliases(command))
-
-
-def _runtime_command_evidence_aliases(command: str) -> tuple[str, ...]:
-    """Return exact runtime command aliases for sibling reconciliation."""
-    aliases = [_normalize_exact_command(command)]
-    single_inner_command = _single_exact_command_after_safe_shell_preamble(command)
-    if single_inner_command and single_inner_command not in aliases:
-        aliases.append(single_inner_command)
-    return tuple(alias for alias in aliases if alias)
-
-
-def _single_exact_command_after_safe_shell_preamble(command: str) -> str | None:
-    """Return one wrapped inner command without lowercasing exact evidence."""
-    body = _shell_command_body(command)
-    if body is None:
-        return None
-    segments = tuple(_segments_after_safe_shell_preamble(body))
-    if len(segments) != 1:
-        return None
-    return _normalize_exact_command(segments[0])
-
-
-def _typed_evidence_is_usable_for_sibling_reconciliation(result: ACExecutionResult) -> bool:
-    """Return True when typed evidence was not rejected by validation/verifier."""
-    if result.typed_evidence is None:
-        return False
-    if result.typed_evidence_error:
-        return False
-    if result.typed_evidence_validation is not None and not result.typed_evidence_validation.ok:
-        return False
-    return result.atomic_verifier_verdict is None or result.atomic_verifier_verdict.passed
-
-
-def _typed_file_evidence_proves_current_existence(
-    result: ACExecutionResult,
-    relative_path: str,
-) -> bool:
-    """Return True when typed file evidence is backed by current end-state existence."""
-    if result.runtime_handle is None or result.runtime_handle.cwd is None:
-        return False
-    candidate = (Path(result.runtime_handle.cwd).resolve() / relative_path).resolve()
-    try:
-        candidate.relative_to(Path(result.runtime_handle.cwd).resolve())
-    except ValueError:
-        return False
-    return candidate.is_file()
-
-
-def _evidence_values_from_result(result: ACExecutionResult) -> tuple[set[str], set[str], set[str]]:
-    """Return normalized file paths, run commands, and passed commands.
-
-    This intentionally uses only structured/runtime evidence, not broad natural
-    language success claims, so sibling AC completion remains conservative.
-    """
-    files: set[str] = set()
-    run_commands: set[str] = set()
-    passed_commands: set[str] = set()
-
-    if _typed_evidence_is_usable_for_sibling_reconciliation(result):
-        assert result.typed_evidence is not None
-        task_cwd = result.runtime_handle.cwd if result.runtime_handle is not None else None
-        for value in _flatten_evidence_values(result.typed_evidence.get("files_touched")):
-            normalized = (
-                _workspace_relative_file_claim(str(value), task_cwd=task_cwd) or str(value).strip()
-            )
-            if normalized and _typed_file_evidence_proves_current_existence(result, normalized):
-                files.add(normalized)
-        for value in _flatten_evidence_values(result.typed_evidence.get("commands_run")):
-            _add_runtime_command_evidence(run_commands, str(value))
-        for value in _flatten_evidence_values(result.typed_evidence.get("tests_passed")):
-            _add_runtime_command_evidence(passed_commands, str(value))
-            _add_runtime_command_evidence(run_commands, str(value))
-
-    for message in result.messages:
-        if not message.tool_name:
-            continue
-
-        if message.tool_name == "Bash":
-            for command in _runtime_message_command_values(message):
-                _add_runtime_command_evidence(run_commands, command)
-            continue
-
-    passed_commands.update(_successful_runtime_test_commands(result.messages))
-    return files, run_commands, passed_commands
-
-
-def _criterion_satisfied_by_evidence(
-    criterion: str,
-    files: set[str],
-    run_commands: set[str],
-    passed_commands: set[str] | None = None,
-) -> bool:
-    """Conservatively decide whether evidence satisfies a sibling criterion."""
-    normalized_run_commands = {
-        _normalize_exact_command(command) for command in run_commands if command
-    }
-    normalized_passed_commands = {
-        _normalize_exact_command(command) for command in (passed_commands or set()) if command
-    }
-
-    for file_path in files:
-        if file_path and _criterion_is_exact_file_presence_ac(criterion, file_path):
-            return True
-
-    for command in normalized_passed_commands:
-        if command and _criterion_is_exact_command_pass_ac(criterion, command):
-            return True
-
-    for command in normalized_run_commands:
-        if command and _criterion_is_exact_command_run_ac(criterion, command):
-            return True
-
-    return False
-
-
-def _criterion_is_exact_file_presence_ac(criterion: str, file_path: str) -> bool:
-    """Return True when the criterion is only an exact file-presence AC."""
-    normalized_path = Path(file_path.strip()).as_posix()
-    if (
-        not normalized_path
-        or Path(normalized_path).is_absolute()
-        or ".." in Path(normalized_path).parts
-    ):
-        return False
-    inline_code_paths = [
-        Path(match.group(1).strip()).as_posix()
-        for match in re.finditer(r"`([^`]+)`", criterion)
-        if match.group(1).strip()
-    ]
-    if inline_code_paths != [normalized_path]:
-        return False
-
-    normalized = _normalized_evidence_text(
-        re.sub(r"`[^`]+`", "<path>", criterion).strip().rstrip(".")
-    )
-    return normalized in {
-        "<path> exists",
-        "<path> is present",
-        "the file <path> exists",
-        "the file <path> is present",
-    }
-
-
-def _criterion_inline_code_values(criterion: str) -> list[str]:
-    """Return normalized inline-code fragments from a criterion."""
-    return [
-        _normalize_exact_command(match.group(1).strip())
-        for match in re.finditer(r"`([^`]+)`", criterion)
-        if match.group(1).strip()
-    ]
-
-
-def _criterion_is_exact_command_pass_ac(criterion: str, command: str) -> bool:
-    """Return True when the criterion is only an exact command-pass AC."""
-    normalized_command = _normalize_exact_command(command)
-    if not normalized_command or _criterion_inline_code_values(criterion) != [normalized_command]:
-        return False
-    normalized = _normalized_evidence_text(
-        re.sub(r"`[^`]+`", "<command>", criterion).strip().rstrip(".")
-    )
-    return normalized in {
-        "<command> passes",
-        "<command> passed",
-        "<command> succeeds",
-        "<command> succeeded",
-        "the exact command <command> passes",
-        "the exact command <command> passed",
-        "the exact command <command> succeeds",
-        "the exact command <command> succeeded",
-        "the exact command <command> exits with code 0",
-        "the command <command> passes",
-        "the command <command> succeeds",
-    }
-
-
-def _criterion_is_exact_command_run_ac(criterion: str, command: str) -> bool:
-    """Return True when the criterion is only an exact command-run AC."""
-    normalized_command = _normalize_exact_command(command)
-    if not normalized_command or _criterion_inline_code_values(criterion) != [normalized_command]:
-        return False
-    normalized = _normalized_evidence_text(
-        re.sub(r"`[^`]+`", "<command>", criterion).strip().rstrip(".")
-    )
-    return normalized in {
-        "run <command>",
-        "run the exact command <command>",
-        "execute <command>",
-        "execute the exact command <command>",
-        "the exact command <command> runs",
-        "the command <command> runs",
-    }
-
-
-def _complete_sibling_acs_from_evidence(
-    *,
-    level_results: list[ACExecutionResult],
-    ac_statuses: dict[int, str],
-    failed_indices: set[int],
-    completed_count: int,
-    level_success: int,
-    level_failed: int,
-) -> tuple[int, int, int, list[ACExecutionResult]]:
-    """Mark sibling ACs satisfied when a successful sibling has exact evidence.
-
-    Observation jobs often ask for separate ACs such as "file exists", "test
-    file exists", and "pytest passes". A single worker can legitimately create
-    both files and run the test. This function reconciles those concrete sibling
-    ACs from runtime/typed evidence instead of leaving them failed or pending.
-    """
-    replacements: dict[int, ACExecutionResult] = {}
-    successful_evidence = [
-        (result.ac_index, *_evidence_values_from_result(result))
-        for result in level_results
-        if result.success
-    ]
-    if not successful_evidence:
-        return completed_count, level_success, level_failed, level_results
-
-    for result in level_results:
-        if result.success or result.outcome != ACExecutionOutcome.FAILED:
-            continue
-        for source_idx, files, run_commands, passed_commands in successful_evidence:
-            if source_idx == result.ac_index:
-                continue
-            if not _criterion_satisfied_by_evidence(
-                result.ac_content,
-                files,
-                run_commands,
-                passed_commands,
-            ):
-                continue
-            replacements[result.ac_index] = ACExecutionResult(
-                ac_index=result.ac_index,
-                ac_content=result.ac_content,
-                success=True,
-                final_message=(f"Satisfied by runtime evidence from sibling AC {source_idx + 1}."),
-                retry_attempt=result.retry_attempt,
-                outcome=ACExecutionOutcome.SATISFIED_EXTERNALLY,
-            )
-            break
-
-    if not replacements:
-        return completed_count, level_success, level_failed, level_results
-
-    reconciled: list[ACExecutionResult] = []
-    for result in level_results:
-        replacement = replacements.get(result.ac_index)
-        if replacement is None:
-            reconciled.append(result)
-            continue
-        if result.outcome == ACExecutionOutcome.FAILED:
-            failed_indices.discard(result.ac_index)
-            if level_failed > 0:
-                level_failed -= 1
-        ac_statuses[result.ac_index] = "completed"
-        completed_count += 1
-        level_success += 1
-        reconciled.append(replacement)
-
-    return completed_count, level_success, level_failed, reconciled
-
-
-def _render_ac_section(
-    result: ACExecutionResult,
-    *,
-    index_path: tuple[int, ...],
-    heading_level: int,
-    include_header: bool = True,
-) -> list[str]:
-    """Render a single Task or Subtask section for verification/audit output."""
-    lines: list[str] = []
-    if include_header:
-        status = "COMPLETED" if result.success else "FAILED"
-        label = "Task" if len(index_path) == 1 else "Subtask"
-        lines.append(
-            f"{'#' * heading_level} {label} {'.'.join(str(i) for i in index_path)}: "
-            f"[{status}] {result.ac_content}"
-        )
-
-    if result.is_decomposed and result.sub_results:
-        lines.append(f"Decomposed into {len(result.sub_results)} Subtasks")
-        for idx, sub_result in enumerate(result.sub_results, start=1):
-            if lines:
-                lines.append("")
-            lines.extend(
-                _render_ac_section(
-                    sub_result,
-                    index_path=index_path + (idx,),
-                    heading_level=min(heading_level + 1, 6),
-                )
-            )
-        return lines
-
-    evidence_lines = _extract_leaf_evidence_lines(result)
-    if evidence_lines:
-        lines.extend(evidence_lines)
-    else:
-        lines.append("Result:")
-        lines.append("No final result message captured.")
-    return lines
 
 
 def _collect_decomposition_depth_warning_paths(
@@ -6102,102 +4089,14 @@ Files present:
         typed_evidence: EvidenceRecord,
         ac_content: str,
     ) -> VerifierVerdict:
-        """Verify leaf evidence is backed by runtime transcript events.
-
-        The verifier deliberately ignores the final result message so the
-        accepted evidence cannot be supported only by the leaf's self-report.
-        """
-        support_messages = tuple(messages[:-1] if messages and messages[-1].is_final else messages)
-        if not support_messages:
-            return VerifierVerdict(
-                passed=False,
-                reasons=("no runtime transcript evidence supports the typed evidence claims",),
-                failure_class="EVIDENCE_MISSING",
-            )
-
-        unsupported: list[str] = []
-        evidence_form_mismatches: list[str] = []
-        backed_commands = tuple(
-            command
-            for command in _flatten_evidence_values(typed_evidence.get("commands_run"))
-            if _runtime_messages_support_command_claim(
-                command,
-                _runtime_support_messages_for_field("commands_run", support_messages),
-            )
+        return _verify_atomic_evidence_against_runtime_messages(
+            messages=messages,
+            typed_evidence=typed_evidence,
+            ac_content=ac_content,
+            execution_profile=self._execution_profile,
+            task_cwd=self._task_cwd,
+            adapter_working_directory=self._adapter.working_directory,
         )
-        effective_schema = _effective_evidence_schema_for_ac(self._execution_profile, ac_content)
-        required_fields = set(effective_schema.required)
-        fields_to_verify = list(effective_schema.required)
-
-        for field_name in fields_to_verify:
-            values = tuple(_flatten_evidence_values(typed_evidence.get(field_name)))
-            if not values:
-                if field_name in required_fields:
-                    unsupported.append(f"{field_name}: no concrete claim values")
-                continue
-            field_messages = _runtime_support_messages_for_field(field_name, support_messages)
-            for value in values:
-                if field_name == "commands_run":
-                    if _runtime_messages_support_command_claim(value, field_messages):
-                        continue
-                    if _runtime_messages_have_masked_test_command_form(
-                        value,
-                        field_messages,
-                    ):
-                        evidence_form_mismatches.append(f"{field_name}: {value}")
-                        unsupported.append(f"{field_name}: {value}")
-                        continue
-                    unsupported.append(f"{field_name}: {value}")
-                    continue
-                if field_name == "files_touched":
-                    if _runtime_messages_support_file_claim(
-                        value,
-                        field_messages,
-                        task_cwd=self._task_cwd or self._adapter.working_directory,
-                    ):
-                        continue
-                    unsupported.append(f"{field_name}: {value}")
-                    continue
-                if field_name == "tests_passed":
-                    if _runtime_messages_support_test_claim(
-                        value=value,
-                        backed_commands=backed_commands,
-                        messages=support_messages,
-                        task_cwd=self._task_cwd or self._adapter.working_directory,
-                    ):
-                        continue
-                    if _runtime_messages_have_masked_test_command_for_test_claim(
-                        value=value,
-                        messages=support_messages,
-                        task_cwd=self._task_cwd or self._adapter.working_directory,
-                    ):
-                        evidence_form_mismatches.append(f"{field_name}: {value}")
-                        unsupported.append(f"{field_name}: {value}")
-                        continue
-                    unsupported.append(f"{field_name}: {value}")
-                    continue
-                if not _runtime_messages_support_claim(value, field_messages):
-                    unsupported.append(f"{field_name}: {value}")
-
-        if unsupported:
-            failure_class = (
-                "EVIDENCE_FORM_MISMATCH"
-                if evidence_form_mismatches and len(evidence_form_mismatches) == len(unsupported)
-                else "FABRICATION_SUSPECTED"
-            )
-            reason_prefix = (
-                "evidence form mismatch; unprotected output-filter pipeline "
-                "cannot prove a clean command claim"
-                if failure_class == "EVIDENCE_FORM_MISMATCH"
-                else "unsupported evidence claims"
-            )
-            return VerifierVerdict(
-                passed=False,
-                reasons=(reason_prefix + ": " + "; ".join(unsupported),),
-                failure_class=failure_class,
-            )
-
-        return VerifierVerdict(passed=True)
 
     async def _emit_atomic_typed_evidence_event(
         self,
